pridani vzoroveho testu, pro studenty se musi vytvorit instance testu...

+++ app/Http/Controllers/AdminController.php
@@ -21,6 +21,11 @@ class AdminController extends Controller
 
     public function delete($id)
     {
+        $user = User::findOrFail($id);
+        if($user->admin){
+            return redirect('admin');
+        }
+
         User::destroy($id);
         return redirect('admin');
     }
